# swagger/src/main/resources/swagger-cost-seiko.yaml
openapi: 3.0.3
info:
  title: Cost seiko
  description: |-
    Cost Seiko API
  contact:
    email: khalil@lokalix-solutions.com
  version: 1.0.0
servers:
  - url: http://localhost:9090
...
paths:
...
  /cost-requests/search:
    post:
      tags:
        - cost-request
      summary: |
        Search cost requests with filter
        Return paginated data ordered by creationDate DESC (includes lines)
      operationId: searchCostRequests
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/archivedFilterParam'
      requestBody:
        description: search cost requests
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRequestSearch'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostRequestsPaginated"
  /cost-requests/engineering/search:
    post:
      tags:
        - cost-request
      summary: |
        Search cost requests with filter
        Return paginated data ordered by creationDate DESC (includes lines)
      operationId: searchEngineeringCostRequests
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/archivedFilterParam'
      requestBody:
        description: search cost requests
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRequestSearch'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostRequestsPaginated"
  /cost-requests/{uid}:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - cost-request
      summary: Retrieve a cost request (includes lines)
      operationId: retrieveCostRequest
      responses:
        '200':
          description: cost request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRequest'
    put:
      tags:
        - cost-request
      summary: Update a cost request
      operationId: updateCostRequest
      requestBody:
        description: update cost request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRequestUpdate'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
  /cost-requests:
    post:
      tags:
        - cost-request
      summary: Create a new cost request
      operationId: createCostRequest
      requestBody:
        description: Create cost request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRequestCreate'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
  /cost-requests/{uid}/validate-for-review:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: validate a cost request and mark it as ready for review
      operationId: validateCostRequestForReview
      responses:
        '200':
          description: ok
  /cost-requests/{uid}/validate-for-estimation:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: validate a cost request under review and mark it as ready for estimation
      operationId: validateCostRequestForEstimation
      responses:
        '200':
          description: ok
  /cost-requests/{uid}/clone:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: Clone a cost request with selected lines
      operationId: cloneCostRequest
      requestBody:
        description: Line UIDs to clone with the cost request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRequestClone'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
  /cost-requests/{uid}/archive:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: Archive a cost request (only if status is ACTIVE/WON/LOST or ABORTED)
      operationId: archiveCostRequest
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
  /cost-requests/{uid}/abort:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: abort a cost request and freeze all related data
      operationId: abortCostRequest
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
  /cost-requests/{uid}/new-revision:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: Create a new revision of a cost request (only if status is ESTIMATED)
      operationId: createNewRevisionCostRequest
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /cost-requests/{uid}/standard-bom/import:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: upload a cost request standard BOM
      operationId: uploadCostRequestStandardBom
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadFiles"
      responses:
        '200':
          description: Successful operation
  /cost-requests/{uid}/custom-bom/import:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: upload a cost request custom BOM with configuration
      operationId: uploadCostRequestCustomBom
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomBomImportBody"
      responses:
        '200':
          description: Successful operation
  /cost-requests/{uid}/files/upload:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: upload a cost request files
      operationId: uploadCostRequestFiles
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadFiles"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileInfo"
  /cost-requests/{uid}/files/download:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: download a cost request files
      operationId: downloadCostRequestFiles
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Uid"
      responses:
        '200':
          description: Successful operation
  /cost-requests/{uid}/files/delete:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: delete a cost request files
      operationId: deleteCostRequestFiles
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Uid"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileInfo"
  /cost-requests/{uid}/files:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - cost-request
      summary: get cost request files
      operationId: getAllCostRequestFiles
      parameters:
        - $ref: '#/components/parameters/uidParam'
      responses:
        '200':
          description: successful operation, all cost request files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileInfo'
  /cost-requests/{uid}/generate-quotation-pdf:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: generate cost request pdf
      operationId: generateCostRequestPdf
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateQuotationPdfBody'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: 'string'
                format: 'byte'
  /cost-requests/{uid}/download-quotation-pdf:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: download cost request pdf
      operationId: downloadCostRequestPdf
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: 'string'
                format: 'byte'
  /cost-requests/{uid}/substitute-material-comment:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - cost-request
      operationId: getCostRequestSubstituteMaterialsComment
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubstituteMaterialsComment'
  /cost-requests/{uid}/approved-by-customer:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary:
      operationId: costRequestApprovedByCustomer
      responses:
        '200':
          description: ok
  /cost-requests/{uid}/rejected-by-customer:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary:
      operationId: costRequestRejectedByCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectBody'
        required: true
      responses:
        '200':
          description: ok
  /cost-requests/{uid}/new-revision-created:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary:
      operationId: createdNewRevisionOfCostRequest
      responses:
        '200':
          description: ok
  /cost-requests/{uid}/extend-expiration:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request
      summary: Extend the expiration date of a cost request
      operationId: extendCostRequestExpiration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRequestExtendExpiration'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedRequest'
        '403':
          $ref: '#/components/responses/ForbiddenRequest'
...
  /cost-requests/{uid}/messages/{messageUid}/delete:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/messageUidParam'
    post:
      tags:
        - cost-request
      summary: delete cost request message
      operationId: deleteCostRequestMessage
      responses:
        '200':
          description: Successful operation, cost request message is deleted
...
components:
  parameters:
    uidParam:
      in: path
      name: uid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
    offsetParam:
      in: query
      name: offset
      required: true
      schema:
        type: integer
        minimum: 0
        default: 0
      description: The number of items to skip before starting to collect the result set.
    limitParam:
      in: query
      name: limit
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
      description: The numbers of items to return.
...
    archivedFilterParam:
      in: query
      name: archivedFilter
      required: true
      schema:
        $ref: '#/components/schemas/ArchivedFilter'
      description: The archived filter
...
  schemas:
...
    Uid:
      type: string
      format: uuid
...
    CostRequestsPaginated:
      type: object
      allOf:
        - $ref: "#/components/schemas/PaginationInfos"
        - type: object
          required:
            - results
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/CostRequest"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

# 2) Apply the security globally to all operations
security:
  - bearerAuth: [ ]         # use the same name as above